final commit

+++ readA.py
@@ -3,6 +3,6 @@ import numpy as np
 
 def readMatrixA(R):
     print("Enter the entries in a single line (separated by space): ")
-    entries = list(map(int, input().split()))
+    entries = list(map(float, input().split()))
     A = np.array(entries).reshape(R, R)
     return A
