+++ readB.py
@@ -5,5 +5,5 @@ def readB():
     List = input(
         "Enter a list separeted by comma as B (it must have same number of rows as A) "
     )
-    B = [int(i) for i in List.split(",")]
+    B = [float(i) for i in List.split(",")]
     return B
